#19 :hammer: Repo 하나로 합쳐서 구현함
