+++ app/src/main/java/project/seo/pictureviewer/data/PictureRepository.kt
@@ -2,6 +2,7 @@ package project.seo.pictureviewer.data
 
 import androidx.paging.Pager
 
-interface MainRepository {
+interface PictureRepository {
     fun fetchPictures(): Pager<Int, Picture>
+    suspend fun getDetail(id: Int): Picture?
 }
