+++ app/src/main/java/project/seo/pictureviewer/di/RepositoryModule.kt
@@ -4,17 +4,12 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import project.seo.pictureviewer.data.DetailRepository
-import project.seo.pictureviewer.data.DetailRepositoryImpl
-import project.seo.pictureviewer.data.MainRepository
-import project.seo.pictureviewer.data.MainRepositoryImpl
+import project.seo.pictureviewer.data.PictureRepository
+import project.seo.pictureviewer.data.PictureRepositoryImpl
 
 @InstallIn(SingletonComponent::class)
 @Module
 abstract class RepositoryModule {
     @Binds
-    abstract fun bindRepository(mainRepositoryImpl: MainRepositoryImpl): MainRepository
-
-    @Binds
-    abstract fun bindDetailRepository(detailRepository: DetailRepositoryImpl): DetailRepository
+    abstract fun bindRepository(mainRepositoryImpl: PictureRepositoryImpl): PictureRepository
 }
